latex doc fo hopfield added
