+++ docs/latex_tutorial/latex_tutorials.tex~
@@ -1,0 +1,78 @@
+%LaTeX book template from Wiley
+\documentclass{wileysev}
+\usepackage{w-bookps}
+%To get started we need to load up the TikZ package, the ‘shapes.geometric’ TikZ library and the ‘arrows’ library.
+\usepackage{tikz}
+\usetikzlibrary{arrows}
+\usetikzlibrary{shapes.geometric}
+
+
+%Begin the document here
+\begin{document}
+
+%Main title for this book
+\booktitle{Flowchart}
+
+\authors{Mageswaran\\
+\affil{Aja}
+}
+
+%Footer information
+\offprintinfo{Latex Tutorials, First Edition}{Mageswaran}
+
+%Bring the titlepage in the output
+\titlepage
+
+%Create the chapter with author details
+\chapter{Flowchart Tutorial}
+\chapterauthors{Mageswaran.D
+\chapteraffil{Aja Team}
+}
+
+%Sub heading in the chapter
+\section{Basic Flowchart}
+
+Reference: https://www.sharelatex.com/blog/2013/08/29/tikz-series-pt3.html
+
+% Create Tikz style, something like typedef in C, where we can specify the shape, color, size, text details etc
+
+\tikzstyle{startstop} = [rectangle, rounded corners, minimum width=3cm, minimum height=1cm,text centered, draw=black, fill=red!30]
+\tikzstyle{io} = [trapezium, trapezium left angle=70, trapezium right angle=110, minimum width=3cm, minimum height=1cm, text centered, draw=black, fill=blue!30]
+%\tikzstyle{process} = [rectangle, minimum width=3cm, minimum height=1cm, text centered, draw=black, fill=orange!30]
+\tikzstyle{process} = [rectangle, minimum width=3cm, minimum height=1cm, text centered, text width=3cm, draw=black, fill=orange!30]
+\tikzstyle{decision} = [diamond, minimum width=3cm, minimum height=1cm, text centered, draw=black, fill=green!30]
+\tikzstyle{arrow} = [thick,->,>=stealth]
+
+\begin{figure}[h!] %Create figure holder
+\begin{tikzpicture}[node distance=2cm] %use the ‘tikzpicture’ environment
+
+% Nodes are very powerful as we can easily position them, make them draw a shape, heavily format them and give them some text. In square brackets at the end of the begin command we specify a node distance of 2cm. This is so that the nodes we use to build the blocks are automatically spaced 2cm apart from their centres.
+
+\node (start) [startstop] {Start};
+\node (in1) [io, below of=start] {Input};
+\node (pro1) [process, below of=in1] {Process 1};
+\node (dec1) [decision, below of=pro1] {Decision 1};
+\node (dec1) [decision, below of=pro1, yshift=-0.5cm] {Decision 1};
+\node (pro2a) [process, below of=dec1, yshift=-0.5cm] {Process 2a};
+\node (pro2b) [process, right of=dec1, xshift=2cm] {Process 2b};
+\node (out1) [io, below of=pro2a] {Output};
+\node (stop) [startstop, below of=out1] {Stop};
+
+\draw [arrow] (start) -- (in1);
+\draw [arrow] (in1) -- (pro1);
+\draw [arrow] (pro1) -- (dec1);
+\draw [arrow] (dec1) -- (pro2a);
+\draw [arrow] (dec1) -- (pro2b);
+\draw [arrow] (dec1) -- node[anchor=east] {yes} (pro2a);
+\draw [arrow] (dec1) -- node[anchor=south] {no} (pro2b);
+
+\draw [arrow] (pro2b) |- (pro1);
+\draw [arrow] (pro2a) -- (out1);
+\draw [arrow] (out1) -- (stop);
+
+\node (pro2a) [process, below of=dec1, yshift=-0.5cm] {Process 2a text text text text text text text text text text};
+
+\end{tikzpicture}
+\end{figure}
+
+\end{document}
