+++ docs/neural_network/biological_neurons/neurons.tex
@@ -1,0 +1,68 @@
+%LaTeX book template from Wiley
+\documentclass{wileysev}
+\usepackage{w-bookps}
+%To draw artificail neural netwoeks
+\usepackage{neuralnetwork}
+\usepackage{graphicx}
+\usepackage{tikz}
+\usetikzlibrary{matrix,chains,positioning,decorations.pathreplacing,arrows}
+%To draw flow charts
+\usetikzlibrary{shapes.geometric}
+%To write down math equations
+\usepackage{mathtools}
+%\usepackage{amsmath}
+
+%To get the rectangular box around the figure
+\usepackage{float}
+\floatstyle{boxed}
+\restylefloat{figure}
+
+%Begin the document here
+\begin{document}
+
+%Main title for this book
+\booktitle{Biological Neuron}
+
+\authors{Mageswaran\\
+\affil{Aja}
+}
+
+%Footer information
+\offprintinfo{Biological Neuron, First Edition}{Mageswaran}
+
+%Bring the titlepage in the output
+\titlepage
+
+%Create the chapter with author details
+\chapter{Introduction on Neural Nwtworks}
+\chapterauthors{Mageswaran.D
+\chapteraffil{Aja Team}
+}
+
+%Sub heading in the chapter
+\section{Neural Processing}
+How do you recognize a face in a crowd? How does an economist predict the
+direction of interest rates? Faced with problems like these, the human brain
+uses a web of interconnected processing elements called neurons to process
+information. Each neuron is autonomous and independent; it does its work
+asynchronously, that is, without any synchronization to other events taking
+place. The two problems posed, namely recognizing a face and forecasting
+interest rates, have two important characteristics that distinguish them from
+other problems: First, the problems are complex, that is, you can’t devise a
+simple step-by-step algorithm or precise formula to give you an answer; and
+second, the data provided to resolve the problems is equally complex and may
+be noisy or incomplete. You could have forgotten your glasses when you’re
+trying to recognize that face. The economist may have at his or her disposal
+thousands of pieces of data that may or may not be relevant to his or her
+forecast on the economy and on interest rates. \\*
+The vast processing power inherent in biological neural structures has inspired
+the study of the structure itself for hints on organizing human-made computing
+structures. Artificial neural networks, the subject of this book, covers the way
+to organize synthetic neurons to solve the same kind of difficult, complex
+problems in a similar manner as we think the human brain may. This chapter
+will give you a sampling of the terms and nomenclature used to talk about
+neural networks. These terms will be covered in more depth in the chapters to
+follow.
+
+
+\end{document}
