+++ docs/neural_network/types_nn/types_of_neural_network.tex
@@ -1,0 +1,67 @@
+%LaTeX book template from Wiley
+\documentclass{wileysev}
+\usepackage{w-bookps}
+%To draw artificail neural netwoeks
+\usepackage{graphicx}
+\usepackage{tikz}
+\usetikzlibrary{matrix,chains,positioning,decorations.pathreplacing,arrows}
+%To draw flow charts
+\usetikzlibrary{shapes.geometric}
+%To write down math equations
+\usepackage{mathtools}
+%\usepackage{amsmath}
+
+%To get the rectangular box around the figure
+\usepackage{float}
+\floatstyle{boxed}
+\restylefloat{figure}
+
+\usepackage{tabto}
+\NumTabs{6}
+
+%Begin the document here
+\begin{document}
+
+%Main title for this book
+\booktitle{Types of Neural Networks}
+
+\authors{Mageswaran\\
+\affil{Aja}
+}
+
+%Footer information
+\offprintinfo{Types of Neural Networks, First Edition}{Mageswaran}
+
+%Bring the titlepage in the output
+\titlepage
+
+%Create the chapter with author details
+\chapter{Based on Learning}
+\chapterauthors{Mageswaran.D
+\chapteraffil{Aja Team}
+}
+
+%\newcommand{\itab}[1]{\hspace{0em}\rlap{#1}}
+%\newcommand{\tab} [1]{\hspace{.2\textwidth}\rlap{#1}}
+
+%Sub heading in the chapter
+%\section{Based on Learning}
+
+A network can be subject to supervised or unsupervised learning. The learning would be supervised if external criteria are used and matched by the network output, and if not, the learning is unsupervised. This is one broad way to divide different neural network approaches. Unsupervised approaches are also termed self-organizing. There is more interaction between neurons, typically with feedback and intralayer connections between neurons promoting self-organization. \\*
+Supervised networks are a little more straightforward to conceptualize than unsupervised networks. You apply the inputs to the supervised network along with an expected response, much like the Pavlovian conditioned stimulus and response regimen. You mold the network with stimulus-response pairs. Astock market forecaster may present economic data (the stimulus) along with metrics of stock market performance (the response) to the neural network to the present and attempt to predict the future once training is complete. \\*
+\tab You provide unsupervised networks with only stimulus. You may, for example, want an unsupervised network to correctly classify parts from a conveyor belt into part numbers, providing an image of each part to do the classification (the stimulus). The unsupervised network in this case would act like a look-up memory that is indexed by its contents, or a Content-Addressable-Memory (CAM). \\*
+
+\tikzstyle{rectangular_box} = [rectangle, minimum width=3cm, minimum height=1cm, text centered, text width=3cm, draw=black, fill=black!30]
+
+\begin{figure}[h!] 
+\centering
+\begin{tikzpicture}[node distance=2cm] 
+
+\node (learning_types) [rectangular box, xshift=8cm] {Learning};
+\node (supervised) [rectangular box, below of=learning_types, xshift=2cm] {Supervised};
+\node (unsupervised) [rectangular box, left of=supervised, xshift=8cm] {UnSupervised};
+
+\end{tikzpicture}
+\end{figure}
+
+\end{document}
