+++ example/ann/CMakeLists.txt
@@ -1,4 +1,4 @@
 INCLUDE_DIRECTORIES(../../include)
 
 ADD_EXECUTABLE(hopfield_network hopfield_network.cpp)
-TARGET_LINK_LIBRARIES(hopfield_network LINK_PUBLIC aja_ann)
+#TARGET_LINK_LIBRARIES(hopfield_network LINK_PUBLIC aja_ann)
