+++ example/ann/hopfield_network.cpp
@@ -1,7 +1,88 @@
-#include <aja/ann/neuron.hpp>
-#include <aja/ann/network.hpp>
+//For background theory : aja/docs/neural_network/hopfield/hopfield_network.pdf
+
+#include <stdio.h>
+#include <iostream>
+#include <math.h>
 
 using namespace std;
+
+class neuron
+{
+protected:
+    int activation;
+    friend class network;
+public:
+    // Wkj -> connection weight from j to neuron k
+    int weight_vector[4];
+    neuron()
+    {
+
+    }
+
+    neuron(int *j)
+    {
+        int i;
+        for(i = 0; i < 4; i++)
+        {
+            weight_vector[i]= *(j + i);
+        }
+    }
+
+    int act(int m, int *x)
+    {
+        int i;
+        int a=0;
+        for(i = 0; i < m; i++)
+        {
+            a += x[i] * weight_vector[i];
+        }
+        return a;
+    }
+};
+
+class network
+{
+public:
+    neuron nrn[4];
+    int output[4];
+    int threshold(int k)
+    {
+        if(k >= 0)
+        {
+            return (1);
+        }
+        else
+        {
+            return (0);
+        }
+    }
+    network(int a[4],int b[4],int c[4],int d[4])
+    {
+        nrn[0] = neuron(a) ;
+        nrn[1] = neuron(b) ;
+        nrn[2] = neuron(c) ;
+        nrn[3] = neuron(d) ;
+    }
+    void activation(int *patrn)
+    {
+        int i,j;
+        for(i = 0; i < 4; i++)
+        {
+            for(j = 0; j < 4; j++)
+            {
+                cout << "neuron["<<i<<"].weight_vec["<<j<<"] is "
+                     << nrn[i].weight_vector[j]
+                        << "\n";
+            }
+            nrn[i].activation = nrn[i].act(4,patrn);
+            cout << "activation is " << nrn[i].activation
+                 << "\n";
+            output[i]=threshold(nrn[i].activation);
+            cout << "output value is " << output[i]
+                    << "\n";
+        }
+    }
+};
 
 int main (int argc, char *argv[])
 {
@@ -37,14 +118,14 @@ int main (int argc, char *argv[])
         if (hf_net.output[i] == patrn1[i])
         {
             cout << " pattern= " << patrn1[i]
-                 << " output = " << hf_net.output[i]
-                 << " component matches!\n";
+                    << " output = " << hf_net.output[i]
+                       << " component matches!\n";
         }
         else
         {
             cout << " pattern= " << patrn1[i]
-                 << " output = " << hf_net.output[i]
-                 << " discrepancy occurred!\n";
+                    << " output = " << hf_net.output[i]
+                       << " discrepancy occurred!\n";
         }
     }
     cout << "========================================\n\n";
@@ -58,14 +139,14 @@ int main (int argc, char *argv[])
         if(hf_net.output[i] == patrn2[i])
         {
             cout << " pattern = " << patrn2[i]
-                 << " output = " << hf_net.output[i]
-                 <<" component matches!\n";
+                    << " output = " << hf_net.output[i]
+                       <<" component matches!\n";
         }
         else
         {
             cout << " pattern= " << patrn2[i]
-                 << " output = " << hf_net.output[i]
-                 << " discrepancy occurred!\n";
+                    << " output = " << hf_net.output[i]
+                       << " discrepancy occurred!\n";
         }
     }
 }
