+++ include/aja/ann/neuron.hpp
@@ -9,22 +9,6 @@
 #ifndef AJA_NEURON_H
 #define AJA_NEURON_H
 
-#include <stdio.h>
-#include <iostream>
-//TODO: Replace with vexCL
-#include <math.h>
 
-class neuron
-{
-protected:
-    int activation;
-    friend class network;
-public:
-    // Wkj -> connection weight from j to neuron k
-    int weightv[4];
-    neuron();
-    neuron(int *j) ;
-    int act(int, int*);
-};
 
 #endif
